אקורדיון+ עדכון סך הוצאות + עדכון dropdown עריכת תת קטגוריה+ הצגת תת-קט' אחרי הוספה +מחיקת תת-קט'
פתיחה וסגירה של האקורדיון קטגוריות, בלחיצה על הכפתור.
עדכון סך ההוצאות/הכנסות בתת קטגוריה לאחר הזנה.
עדכון הdropdown של הקטגוריות בעריכת תת קטגוריה, כך שהקטגוריה הרלוונטית תוצג ראשונה.
לאחר הוספת תת קטגוריה היא תופיע אוטומטית בעמוד.
מחיקת תת קטגוריה.

+++ Shared/Models/present/toAdd/SubCategoryToAdd.cs
@@ -11,7 +11,7 @@ namespace CountOnIt.Shared.Models.present.toAdd
         public int id { get; set; }
         public int categoryID { get; set; }
         public string subCategoryTitle { get; set; }
-        public int? monthlyPlannedBudget { get; set; }
+        public double? monthlyPlannedBudget { get; set; }
         public int? importance { get; set; }
     }
 }
